refactor: moved all views in Profile and created Auth page placeholder
