+++ apps/front/src/views/auth/Index.vue
@@ -1,0 +1,13 @@
+<template>
+  <div class="h-full">Password Page W.I.P</div>
+</template>
+
+<script>
+export default {
+  name: 'LnmUmbrelAuth',
+
+  setup() {
+    return {}
+  },
+}
+</script>
